requirements.txt
streamlit>=1.45.1
requests>=2.31.0
langgraph>=0.3.1
langchain>=0.3.1
google-generativeai
# sqlite3
python-dotenv
typing-extensions
